Add custom block load

+++ src/me/lofienjoyer/valkyrie/Chunk.java
@@ -9,7 +9,7 @@ import java.util.concurrent.Future;
 
 public class Chunk {
 
-    private short[] data;
+    private int[] data;
     private final Vector2i position;
     private boolean dirty;
     private boolean priority;
@@ -32,7 +32,7 @@ public class Chunk {
     public void setBlock(int x, int y, int z, int id) {
         if (data == null) return;
         if (y < 0 || y > 127 || x < 0 || x > 31 || z < 0 || z > 31) return;
-        data[x | y << 5 | z << 12] = (short) (id << 12);
+        data[x | y << 5 | z << 12] = (id << 12);
     }
 
     public int getRedLight(int x, int y, int z) {
@@ -44,7 +44,7 @@ public class Chunk {
     public void setRedLight(int x, int y, int z, int light) {
         if (data == null) return;
         if (y < 0 || y > 127 || x < 0 || x > 31 || z < 0 || z > 31) return;
-        data[x | y << 5 | z << 12] = (short) ((data[x | y << 5 | z << 12] & 0xf1ff) | (light << 9));
+        data[x | y << 5 | z << 12] = ((data[x | y << 5 | z << 12] & 0xf1ff) | (light << 9));
     }
 
     public int getGreenLight(int x, int y, int z) {
@@ -56,7 +56,7 @@ public class Chunk {
     public void setGreenLight(int x, int y, int z, int light) {
         if (data == null) return;
         if (y < 0 || y > 127 || x < 0 || x > 31 || z < 0 || z > 31) return;
-        data[x | y << 5 | z << 12] = (short) ((data[x | y << 5 | z << 12] & 0xfe3f) | (light << 6));
+        data[x | y << 5 | z << 12] = ((data[x | y << 5 | z << 12] & 0xfe3f) | (light << 6));
     }
 
     public int getBlueLight(int x, int y, int z) {
@@ -68,7 +68,7 @@ public class Chunk {
     public void setBlueLight(int x, int y, int z, int light) {
         if (data == null) return;
         if (y < 0 || y > 127 || x < 0 || x > 31 || z < 0 || z > 31) return;
-        data[x | y << 5 | z << 12] = (short) ((data[x | y << 5 | z << 12] & 0xffc7) | (light << 3));
+        data[x | y << 5 | z << 12] = ((data[x | y << 5 | z << 12] & 0xffc7) | (light << 3));
     }
 
     public int getSunLight(int x, int y, int z) {
@@ -80,14 +80,14 @@ public class Chunk {
     public void setSunLight(int x, int y, int z, int light) {
         if (data == null) return;
         if (y < 0 || y > 127 || x < 0 || x > 31 || z < 0 || z > 31) return;
-        data[x | y << 5 | z << 12] = (short) ((data[x | y << 5 | z << 12] & 0xfff8) | light);
+        data[x | y << 5 | z << 12] = ((data[x | y << 5 | z << 12] & 0xfff8) | light);
     }
 
-    public short[] getData() {
+    public int[] getData() {
         return data;
     }
 
-    public void setData(short[] data) {
+    public void setData(int[] data) {
         this.data = data;
     }
